Refactor to cleanup code base and fix various bugs.
Refactored the stapler trait, attachment, and storage drivers to fix a
bug when using STAPLER_NULL.  Events now fire listeners on the attachment (instead of inside the
stapler trait).  Deletes and writes are now pushed into queues that are
flushed after a save or delete.  Added listeners for the 'deleting'
event to queue files to be deleted.
Simplified file storage driver, various other minor tweaks and
improvements.